+++ src/nbs/interfaces/layer/LayerMeta.ts
@@ -1,0 +1,16 @@
+/**
+ * Meta information for a {@linkcode Layer}.
+ */
+export default interface LayerMeta {
+    /**
+     * The name of the layer.
+     */
+    name: string
+};
+
+/**
+ * Default {@linkcode LayerMeta} values.
+ */
+export const defaultLayerMeta: LayerMeta = {
+    "name": ""
+};