+++ src/util/util.ts
@@ -1,43 +1,80 @@
-import Instrument from "../nbs/Instrument";
+import Instrument from "../nbs/instrument/Instrument";
 import Note from "../nbs/Note";
 import Layer from "../nbs/Layer";
 
+/**
+ * A buffer object wrapper.
+ */
 class Buffer {
-    public buffer: ArrayBuffer;
-    protected viewer: DataView;
-    public currentByte = 0;
+    /**
+     * The buffer that is being read.
+     */
+    readonly buffer: ArrayBuffer;
 
+    /**
+     * DataView for the buffer.
+     */
+    protected viewer: DataView;
+
+    /**
+     * Next byte to read.
+     */
+    public nextByte = 0;
+
+    /**
+     * Create a buffer wrapper.
+     *
+     * @param buffer ArrayBuffer to read
+     */
     constructor(buffer: ArrayBuffer) {
         this.buffer = buffer;
         this.viewer = new DataView(buffer);
     }
 }
 
+/**
+ * Represents an ArrayBuffer reader.
+ */
 export class BufferReader extends Buffer {
+    /**
+     * Read the next byte.
+     */
     public readByte(): number {
-        const result = this.viewer.getInt8(this.currentByte);
-        this.currentByte += 1;
+        const result = this.viewer.getInt8(this.nextByte);
+        this.nextByte += 1;
         return result;
     }
 
+    /**
+     * Read the next unsigned byte.
+     */
     public readUnsingedByte(): number {
-        const result = this.viewer.getUint8(this.currentByte);
-        this.currentByte += 1;
+        const result = this.viewer.getUint8(this.nextByte);
+        this.nextByte += 1;
         return result;
     }
 
+    /**
+     * Read the next short.
+     */
     public readShort(): number {
-        const result = this.viewer.getInt16(this.currentByte, true);
-        this.currentByte += 2;
+        const result = this.viewer.getInt16(this.nextByte, true);
+        this.nextByte += 2;
         return result;
     }
 
+    /**
+     * Read the next integer.
+     */
     public readInt(): number {
-        const result = this.viewer.getInt32(this.currentByte, true);
-        this.currentByte += 4;
+        const result = this.viewer.getInt32(this.nextByte, true);
+        this.nextByte += 4;
         return result;
     }
 
+    /**
+     * Read the next string.
+     */
     public readString(): string {
         const length = this.readInt();
         let result = "";
@@ -49,59 +86,117 @@ export class BufferReader extends Buffer {
     }
 }
 
+/**
+ * Represents an ArrayBuffer writer.
+ */
 export class BufferWriter extends Buffer {
-    private dry: boolean;
+    /**
+     * Whether to execute a dry run.
+     * Used to find the target size of the buffer.
+     */
+    private readonly dry: boolean;
 
+    /**
+     * Create a buffer writer.
+     *
+     * @param buffer ArrayBuffer to read
+     * @param dry Whether to execute a dry run
+     */
     constructor(buffer: ArrayBuffer, dry = false) {
         super(buffer);
 
         this.dry = dry;
     }
 
-    public writeByte(val: number): void {
+    /**
+     * Write a byte.
+     *
+     * @param val Value to write
+     */
+    public writeByte(val: number | undefined): void {
+        val ??= 0;
+
         if (!this.dry) {
-            this.viewer.setInt8(this.currentByte, Math.floor(val));
+            this.viewer.setInt8(this.nextByte, Math.floor(val));
         }
 
-        this.currentByte += 1;
+        this.nextByte += 1;
     }
 
-    public writeUnsignedByte(val: number): void {
+    /**
+     * Write an unsigned byte.
+     *
+     * @param val Value to write
+     */
+    public writeUnsignedByte(val: number | undefined): void {
+        val ??= 0;
+
         if (!this.dry) {
-            this.viewer.setUint8(this.currentByte, val);
+            this.viewer.setUint8(this.nextByte, val);
         }
 
-        this.currentByte += 1;
+        this.nextByte += 1;
     }
 
-    public writeShort(val: number): void {
+    /**
+     * Write a short.
+     *
+     * @param val Value to write
+     */
+    public writeShort(val: number | undefined): void {
+        val ??= 0;
+
         if (!this.dry) {
-            this.viewer.setInt16(this.currentByte, val, true);
+            this.viewer.setInt16(this.nextByte, val, true);
         }
 
-        this.currentByte += 2;
+        this.nextByte += 2;
     }
 
-    public writeInt(val: number): void {
+    /**
+     * Write an integer.
+     *
+     * @param val Value to write
+     */
+    public writeInt(val: number | undefined): void {
+        val ??= 0;
+
         if (!this.dry) {
-            this.viewer.setInt32(this.currentByte, val, true);
+            this.viewer.setInt32(this.nextByte, val, true);
         }
 
-        this.currentByte += 4;
+        this.nextByte += 4;
     }
 
-    public writeString(val: string): void {
+    /**
+     * Write a string.
+     *
+     * @param val Value to write
+     */
+    public writeString(val: string | undefined): void {
+        val ??= "";
+
         this.writeInt(val.length);
         for (const i of val) {
-            // eslint-disable-next-line unicorn/prefer-code-point
             this.writeUnsignedByte(i.charCodeAt(0));
         }
     }
 }
 
-let LayerClass = Layer;
-let NoteClass = Note;
-let InstrumentClass = Instrument;
+/**
+ * The {@linkcode Layer} to utilize within the API.
+ */
+let LayerClass: Layer;
+
+/**
+ * The {@linkcode Note} to utilize within the API.
+ */
+let NoteClass: Note;
+
+/**
+ * The {@linkcode Instrument} to utilize within the API.
+ */
+let InstrumentClass: Instrument;
 
 /**
  * Get the {@linkcode Layer} class.